+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".authentication.LoginActivity" />
         <activity android:name=".authentication.RegistrationActivity" />
         <activity android:name=".calculator.CalculatorActivity" />
         <activity android:name=".photo_gallery.PhotoActivity" />
