+++ app/src/main/java/com/codecool/arinyu/myfirstandroidapp/MainActivity.java
@@ -1,19 +1,107 @@
 package com.codecool.arinyu.myfirstandroidapp;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
+import android.text.TextUtils;
+import android.view.KeyEvent;
+import android.view.View;
+import android.view.inputmethod.EditorInfo;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
 
+import com.codecool.arinyu.myfirstandroidapp.authentication.RegistrationActivity;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 public class MainActivity extends AppCompatActivity {
+    private EditText mPassword, mEmailAddress;
+    private FirebaseAuth mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_first);
+
+        mAuth = FirebaseAuth.getInstance();
+        mEmailAddress = (EditText) findViewById(R.id.txtEmailAddress);
+        mPassword = (EditText) findViewById(R.id.txtPassword);
+
+        final Button loginButton = (Button) findViewById(R.id.loginButton);
+        final Button signupButton = (Button) findViewById(R.id.signupButton);
+        final ProgressBar loginProgress = (ProgressBar) findViewById(R.id.login_progress);
+
+        mPassword.setOnEditorActionListener(new EditText.OnEditorActionListener() {
+            @Override
+            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
+                if (actionId == EditorInfo.IME_ACTION_DONE) {
+                    loginButton.performClick();
+                    return true;
+                }
+                return false;
+            }
+        });
+
+        loginButton.setOnClickListener(new Button.OnClickListener() {
+            public void onClick(View v) {
+                String email = mEmailAddress.getText().toString();
+                String password = mPassword.getText().toString();
+                if (TextUtils.isEmpty(email)) {
+                    Toast.makeText(getApplicationContext(), "Enter email address!", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+
+                if (TextUtils.isEmpty(password)) {
+                    Toast.makeText(getApplicationContext(), "Enter password!", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+
+                mAuth.signInWithEmailAndPassword(email, password)
+                        .addOnCompleteListener(MainActivity.this, new OnCompleteListener<AuthResult>() {
+                            @Override
+                            public void onComplete(@NonNull Task<AuthResult> task) {
+                                loginProgress.setVisibility(View.GONE);
+                                if (task.isSuccessful()) {
+                                    // Sign in success
+                                    Intent billSplitterIntent = new Intent(MainActivity.this, BillSplitterActivity.class);
+                                    startActivity(billSplitterIntent);
+                                    Toast.makeText(MainActivity.this, "You logged in successfully :)",
+                                            Toast.LENGTH_SHORT).show();
+                                    finish();
+
+                                } else {
+                                    // If sign in fails, display a message to the user.
+                                    Toast.makeText(MainActivity.this, "Authentication failed!",
+                                            Toast.LENGTH_SHORT).show();
+
+                                    Toast.makeText(MainActivity.this, task.getException().getMessage(),
+                                            Toast.LENGTH_LONG).show();
+                                    FirebaseUser user = mAuth.getCurrentUser();
+                                    user = null;
+                                }
+                            }
+                        });
+            }
+        });
+        signupButton.setOnClickListener(new Button.OnClickListener() {
+            public void onClick(View v) {
+                Intent registerIntent = new Intent(MainActivity.this, RegistrationActivity.class);
+                startActivity(registerIntent);
+            }
+        });
     }
+
     @Override
     public void onBackPressed() {
-
-            super.onBackPressed();
+        Intent mainIntent = new Intent(this, MainActivity.class);
+        startActivity(mainIntent);
+        super.onBackPressed();
     }
 }
