--- app/src/main/java/com/codecool/arinyu/myfirstandroidapp/authentication/LoginActivity.java
@@ -1,104 +1,0 @@
-package com.codecool.arinyu.myfirstandroidapp.authentication;
-
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.text.TextUtils;
-import android.util.Log;
-import android.view.KeyEvent;
-import android.view.View;
-import android.view.inputmethod.EditorInfo;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.codecool.arinyu.myfirstandroidapp.MainActivity;
-import com.codecool.arinyu.myfirstandroidapp.R;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-
-public class LoginActivity extends AppCompatActivity {
-
-    private EditText mPassword, mEmailAddress;
-    private FirebaseAuth mAuth;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-//        if (mAuth.getCurrentUser() != null) {
-//            startActivity(new Intent(this, MainActivity.class));
-//            finish();
-//        }
-
-        setContentView(R.layout.activity_login);
-
-        mAuth = FirebaseAuth.getInstance();
-        mEmailAddress = (EditText) findViewById(R.id.txtEmailAddress);
-        mPassword = (EditText) findViewById(R.id.txtPassword);
-
-
-        final Button btnLogin = (Button) findViewById(R.id.email_sign_in_button);
-        final ProgressBar loginProgress = (ProgressBar) findViewById(R.id.login_progress);
-
-        mPassword.setOnEditorActionListener(new EditText.OnEditorActionListener() {
-            @Override
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                if (actionId == EditorInfo.IME_ACTION_DONE) {
-                    btnLogin.performClick();
-                    return true;
-                }
-                return false;
-            }
-        });
-
-        btnLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String email = mEmailAddress.getText().toString();
-                String password = mPassword.getText().toString();
-
-                if (TextUtils.isEmpty(email)) {
-                    Toast.makeText(getApplicationContext(), "Enter email address!", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-
-                if (TextUtils.isEmpty(password)) {
-                    Toast.makeText(getApplicationContext(), "Enter password!", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-
-                loginProgress.setVisibility(View.VISIBLE);
-
-                // Compare data with Firebase
-                mAuth.signInWithEmailAndPassword(email, password)
-                        .addOnCompleteListener(LoginActivity.this, new OnCompleteListener<AuthResult>() {
-                            @Override
-                            public void onComplete(@NonNull Task<AuthResult> task) {
-                                loginProgress.setVisibility(View.GONE);
-                                if (task.isSuccessful()) {
-                                    // Sign in success, update UI with the signed-in user's information
-                                    Intent intent = new Intent(LoginActivity.this, MainActivity.class);
-                                    startActivity(intent);
-                                    finish();
-
-                                } else {
-                                    // If sign in fails, display a message to the user.
-                                    Toast.makeText(LoginActivity.this, "Authentication failed.",
-                                            Toast.LENGTH_SHORT).show();
-                                    FirebaseUser user = mAuth.getCurrentUser();
-                                    user = null;
-                                }
-                            }
-                        });
-            }
-        });
-    }
-}
-
