+++ app/src/main/res/layout/activity_login.xml
@@ -8,8 +8,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".authentication.LoginActivity">
+    android:paddingTop="@dimen/activity_vertical_margin">
 
     <!-- Login progress -->
     <ProgressBar
@@ -67,7 +66,7 @@
             </android.support.design.widget.TextInputLayout>
 
             <Button
-                android:id="@+id/email_sign_in_button"
+                android:id="@+id/loginButton"
                 style="?android:textAppearanceSmall"
                 android:layout_width="100dp"
                 android:layout_height="wrap_content"
