+++ app/src/main/res/layout/activity_registration.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="#6B4E71"
     android:gravity="center_horizontal"
     android:orientation="vertical"
     android:paddingBottom="@dimen/activity_vertical_margin"
@@ -104,7 +105,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="16dp"
                 android:text="@string/signUp"
-                android:textStyle="bold" />
+                android:textStyle="bold"
+                android:background="@drawable/rounded_button"/>
 
         </LinearLayout>
     </ScrollView>
