+++ app/src/main/res/values/styles.xml
@@ -3,8 +3,8 @@
     <!-- Base application theme. -->
     <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
         <!-- Customize your theme here. -->
-        <item name="colorPrimary">@color/colorPrimary</item>
-        <item name="colorPrimaryDark">#2A2B2A</item>
+        <item name="colorPrimary">#2A2B2A</item>
+        <item name="colorPrimaryDark">#000000</item> <!--almost black-->
         <item name="colorAccent">@color/colorAccent</item>
     </style>
